# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NER Smart Logistics & Accessibility Intelligence Platform",
    "description": "Backend API for the Smart Logistics & Accessibility Intelligence Platform for North-East India (SIH 2026 \u2014 Problem 26002, MDoNER). Supports field incident reporting, road network status, routing, fleet tracking, and logistics coordination for humanitarian operations.",
    "contact": {
      "name": "SIH 2026 Team"
    },
    "license": {
      "name": "Restricted \u2014 Pilot Use Only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/routes/evaluate": {
      "post": {
        "tags": [
          "Routing & Dispatch"
        ],
        "summary": "Evaluate Route",
        "operationId": "evaluate_route_api_v1_routes_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/routes/{route_plan_id}": {
      "get": {
        "tags": [
          "Routing & Dispatch"
        ],
        "summary": "Get Route Plan",
        "operationId": "get_route_plan_api_v1_routes__route_plan_id__get",
        "parameters": [
          {
            "name": "route_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Route Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/dispatch-decisions": {
      "post": {
        "tags": [
          "Routing & Dispatch"
        ],
        "summary": "Record Dispatch Decision",
        "operationId": "record_dispatch_decision_api_v1_trips__trip_id__dispatch_decisions_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlternativeRouteResponse": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "total_distance_meters": {
            "type": "integer",
            "title": "Total Distance Meters"
          },
          "total_duration_seconds": {
            "type": "integer",
            "title": "Total Duration Seconds"
          },
          "geometry": {
            "type": "object",
            "title": "Geometry"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/RouteEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "total_distance_meters",
          "total_duration_seconds",
          "geometry",
          "edges"
        ],
        "title": "AlternativeRouteResponse"
      },
...
      "DispatchAction": {
        "type": "string",
        "enum": [
          "ACCEPTED",
          "REJECTED",
          "DIVERTED"
        ],
        "title": "DispatchAction"
      },
      "DispatchDecisionRequest": {
        "properties": {
          "route_plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Route Plan Id"
          },
          "action": {
            "$ref": "#/components/schemas/DispatchAction"
          },
          "selected_alternative_rank": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Selected Alternative Rank",
            "default": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "route_plan_id",
          "action",
          "reason"
        ],
        "title": "DispatchDecisionRequest"
      },
      "DispatchDecisionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "trip_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trip Id"
          },
          "route_plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Route Plan Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "selected_alternative_rank": {
            "type": "integer",
            "title": "Selected Alternative Rank"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status_version_at_decision": {
            "type": "integer",
            "title": "Status Version At Decision"
          },
          "decided_at": {
            "type": "string",
            "format": "date-time",
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trip_id",
          "route_plan_id",
          "action",
          "selected_alternative_rank",
          "reason",
          "status_version_at_decision",
          "decided_at"
        ],
        "title": "DispatchDecisionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyVersion": {
        "type": "string",
        "enum": [
          "CONSERVATIVE_CRITICAL_V1",
          "STANDARD_DISPATCH_V1"
        ],
        "title": "PolicyVersion"
      },
...
      "RouteEdgeResponse": {
        "properties": {
          "edge_id": {
            "type": "string",
            "format": "uuid",
            "title": "Edge Id"
          },
          "sequence_order": {
            "type": "integer",
            "title": "Sequence Order"
          },
          "cumulative_distance_meters": {
            "type": "integer",
            "title": "Cumulative Distance Meters"
          },
          "cumulative_duration_seconds": {
            "type": "integer",
            "title": "Cumulative Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "edge_id",
          "sequence_order",
          "cumulative_distance_meters",
          "cumulative_duration_seconds"
        ],
        "title": "RouteEdgeResponse"
      },
      "RouteEvaluationRequest": {
        "properties": {
          "trip_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "origin_node_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Node Id"
          },
          "origin_lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Lat"
          },
          "origin_lon": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Lon"
          },
          "destination_node_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Node Id"
          },
          "destination_lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Lat"
          },
          "destination_lon": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Lon"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id"
          },
          "max_weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Weight Kg"
          },
          "height_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Height M",
            "default": 3.0
          },
          "is_hazmat": {
            "type": "boolean",
            "title": "Is Hazmat",
            "default": false
          },
          "cargo_priority": {
            "type": "string",
            "title": "Cargo Priority",
            "default": "TIER_3_STANDARD"
          },
          "departure_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Time"
          },
          "policy_version": {
            "$ref": "#/components/schemas/PolicyVersion",
            "default": "CONSERVATIVE_CRITICAL_V1"
          }
        },
        "type": "object",
        "required": [
          "max_weight_kg"
        ],
        "title": "RouteEvaluationRequest"
      },
      "RoutePlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "trip_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "graph_version": {
            "type": "string",
            "title": "Graph Version"
          },
          "status_version": {
            "type": "integer",
            "title": "Status Version"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "result_status": {
            "$ref": "#/components/schemas/RouteResultStatus"
          },
          "total_distance_meters": {
            "type": "integer",
            "title": "Total Distance Meters"
          },
          "total_duration_seconds": {
            "type": "integer",
            "title": "Total Duration Seconds"
          },
          "requires_human_review": {
            "type": "boolean",
            "title": "Requires Human Review"
          },
          "excluded_edge_reasons": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Excluded Edge Reasons"
          },
          "primary_geometry": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Geometry"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/RouteEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AlternativeRouteResponse"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "trip_id",
          "graph_version",
          "status_version",
          "policy_version",
          "result_status",
          "total_distance_meters",
          "total_duration_seconds",
          "requires_human_review",
          "excluded_edge_reasons",
          "primary_geometry",
          "edges",
          "alternatives",
          "evaluated_at",
          "expires_at"
        ],
        "title": "RoutePlanResponse"
      },
      "RouteResultStatus": {
        "type": "string",
        "enum": [
          "FEASIBLE",
          "NO_FEASIBLE_PATH",
          "INSUFFICIENT_DATA"
        ],
        "title": "RouteResultStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}